move NavDestination and NavigationSetup to navigation
