+++ codegen-compiler/src/main/kotlin/com/freeletics/khonshu/codegen/util/External.kt
@@ -32,12 +32,12 @@ internal val navEventNavigator = ClassName("com.freeletics.khonshu.navigation", 
 internal val navigationExecutor = ClassName("com.freeletics.khonshu.navigation.internal", "NavigationExecutor")
 internal val destinationId = MemberName("com.freeletics.khonshu.navigation.internal", "destinationId")
 internal val navHost = MemberName("com.freeletics.khonshu.navigation.compose", "NavHost")
-internal val navigationSetup = MemberName("com.freeletics.khonshu.navigation.compose", "NavigationSetup")
-internal val navigationDestination = ClassName("com.freeletics.khonshu.navigation.compose", "NavDestination")
-internal val screenDestination = MemberName("com.freeletics.khonshu.navigation.compose", "ScreenDestination")
-internal val overlayDestination = MemberName("com.freeletics.khonshu.navigation.compose", "OverlayDestination")
+internal val navigationSetup = MemberName("com.freeletics.khonshu.navigation", "NavigationSetup")
+internal val navigationDestination = ClassName("com.freeletics.khonshu.navigation", "NavDestination")
+internal val screenDestination = MemberName("com.freeletics.khonshu.navigation", "ScreenDestination")
+internal val overlayDestination = MemberName("com.freeletics.khonshu.navigation", "OverlayDestination")
 internal val localNavigationExecutor =
-    MemberName("com.freeletics.khonshu.navigation.compose", "LocalNavigationExecutor")
+    MemberName("com.freeletics.khonshu.navigation", "LocalNavigationExecutor")
 internal val deepLinkHandler = ClassName("com.freeletics.khonshu.navigation.deeplinks", "DeepLinkHandler")
 internal val deepLinkPrefix = deepLinkHandler.nestedClass("Prefix")
 internal val internalNavigatorApi = ClassName("com.freeletics.khonshu.navigation.internal", "InternalNavigationApi")
