+++ navigation-compose/api/navigation-compose.api
@@ -1,11 +1,3 @@
-public final class com/freeletics/khonshu/navigation/compose/ActivityDestination : com/freeletics/khonshu/navigation/compose/NavDestination {
-	public static final field $stable I
-	public synthetic fun <init> (Lkotlin/reflect/KClass;Landroid/content/Intent;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
-}
-
-public abstract interface class com/freeletics/khonshu/navigation/compose/NavDestination {
-}
-
 public final class com/freeletics/khonshu/navigation/compose/NavHostDefaults {
 	public static final field $stable I
 	public static final field INSTANCE Lcom/freeletics/khonshu/navigation/compose/NavHostDefaults;
@@ -31,26 +23,6 @@ public final class com/freeletics/khonshu/navigation/compose/NavHostTransitionAn
 
 public final class com/freeletics/khonshu/navigation/compose/NavHostTransitionAnimations$Companion {
 	public final fun noAnimations ()Lcom/freeletics/khonshu/navigation/compose/NavHostTransitionAnimations;
-}
-
-public final class com/freeletics/khonshu/navigation/compose/NavigationSetupKt {
-	public static final fun NavigationSetup (Lcom/freeletics/khonshu/navigation/NavEventNavigator;Landroidx/compose/runtime/Composer;I)V
-}
-
-public final class com/freeletics/khonshu/navigation/compose/OverlayDestination : com/freeletics/khonshu/navigation/compose/ContentDestination {
-	public static final field $stable I
-	public synthetic fun <init> (Lkotlin/reflect/KClass;Ljava/io/Serializable;Lkotlin/jvm/functions/Function3;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun getContent ()Lkotlin/jvm/functions/Function3;
-	public fun getExtra ()Ljava/io/Serializable;
-	public fun getId-4Y2Adlc ()Lkotlin/reflect/KClass;
-}
-
-public final class com/freeletics/khonshu/navigation/compose/ScreenDestination : com/freeletics/khonshu/navigation/compose/ContentDestination {
-	public static final field $stable I
-	public synthetic fun <init> (Lkotlin/reflect/KClass;Ljava/io/Serializable;Lkotlin/jvm/functions/Function3;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun getContent ()Lkotlin/jvm/functions/Function3;
-	public fun getExtra ()Ljava/io/Serializable;
-	public fun getId-4Y2Adlc ()Lkotlin/reflect/KClass;
 }
 
 public final class com/freeletics/khonshu/navigation/compose/internal/ComposableSingletons$OverlayNavigatorKt {
