+++ navigation-experimental/api/navigation-experimental.api
@@ -1,32 +1,4 @@
-public final class com/freeletics/khonshu/navigation/compose/ActivityDestination : com/freeletics/khonshu/navigation/compose/NavDestination {
-	public static final field $stable I
-	public synthetic fun <init> (Lkotlin/reflect/KClass;Landroid/content/Intent;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
-}
-
-public abstract interface class com/freeletics/khonshu/navigation/compose/NavDestination {
-}
-
 public final class com/freeletics/khonshu/navigation/compose/NavHostKt {
 	public static final fun NavHost (Lcom/freeletics/khonshu/navigation/NavRoot;Lkotlinx/collections/immutable/ImmutableSet;Landroidx/compose/ui/Modifier;Lkotlinx/collections/immutable/ImmutableSet;Lkotlinx/collections/immutable/ImmutableSet;Lcom/freeletics/khonshu/navigation/NavEventNavigator;Lkotlin/jvm/functions/Function1;Landroidx/compose/runtime/Composer;II)V
-}
-
-public final class com/freeletics/khonshu/navigation/compose/NavigationSetupKt {
-	public static final fun NavigationSetup (Lcom/freeletics/khonshu/navigation/NavEventNavigator;Landroidx/compose/runtime/Composer;I)V
-}
-
-public final class com/freeletics/khonshu/navigation/compose/OverlayDestination : com/freeletics/khonshu/navigation/compose/ContentDestination {
-	public static final field $stable I
-	public synthetic fun <init> (Lkotlin/reflect/KClass;Ljava/io/Serializable;Lkotlin/jvm/functions/Function3;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun getContent ()Lkotlin/jvm/functions/Function3;
-	public fun getExtra ()Ljava/io/Serializable;
-	public fun getId-4Y2Adlc ()Lkotlin/reflect/KClass;
-}
-
-public final class com/freeletics/khonshu/navigation/compose/ScreenDestination : com/freeletics/khonshu/navigation/compose/ContentDestination {
-	public static final field $stable I
-	public synthetic fun <init> (Lkotlin/reflect/KClass;Ljava/io/Serializable;Lkotlin/jvm/functions/Function3;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun getContent ()Lkotlin/jvm/functions/Function3;
-	public fun getExtra ()Ljava/io/Serializable;
-	public fun getId-4Y2Adlc ()Lkotlin/reflect/KClass;
 }
