+++ navigation-experimental/src/main/kotlin/com/freeletics/khonshu/navigation/compose/internal/MultiStack.kt
@@ -6,9 +6,9 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.core.os.bundleOf
 import com.freeletics.khonshu.navigation.BaseRoute
+import com.freeletics.khonshu.navigation.ContentDestination
 import com.freeletics.khonshu.navigation.NavRoot
 import com.freeletics.khonshu.navigation.NavRoute
-import com.freeletics.khonshu.navigation.compose.ContentDestination
 import com.freeletics.khonshu.navigation.internal.DestinationId
 import com.freeletics.khonshu.navigation.internal.destinationId
 import java.util.UUID
