+++ navigation/api/android/navigation.api
@@ -6,6 +6,7 @@ public abstract interface class com/freeletics/khonshu/navigation/ActivityResult
 }
 
 public final class com/freeletics/khonshu/navigation/ActivityResultRequest : com/freeletics/khonshu/navigation/ContractResultOwner {
+	public static final field $stable I
 }
 
 public abstract interface class com/freeletics/khonshu/navigation/ActivityRoute : android/os/Parcelable {
@@ -26,18 +27,32 @@ public abstract interface class com/freeletics/khonshu/navigation/BaseRoute : an
 }
 
 public abstract class com/freeletics/khonshu/navigation/ContractResultOwner : com/freeletics/khonshu/navigation/ResultOwner {
+	public static final field $stable I
 }
 
 public abstract interface class com/freeletics/khonshu/navigation/ExternalActivityRoute : com/freeletics/khonshu/navigation/ActivityRoute {
 	public fun fillInIntent ()Landroid/content/Intent;
 }
 
+public final class com/freeletics/khonshu/navigation/InitialValue$Creator : android/os/Parcelable$Creator {
+	public fun <init> ()V
+	public final fun createFromParcel (Landroid/os/Parcel;)Lcom/freeletics/khonshu/navigation/InitialValue;
+	public synthetic fun createFromParcel (Landroid/os/Parcel;)Ljava/lang/Object;
+	public final fun newArray (I)[Lcom/freeletics/khonshu/navigation/InitialValue;
+	public synthetic fun newArray (I)[Ljava/lang/Object;
+}
+
 public abstract class com/freeletics/khonshu/navigation/InternalActivityRoute : com/freeletics/khonshu/navigation/ActivityRoute {
+	public static final field $stable I
 	public fun <init> ()V
 	public final fun fillInIntent ()Landroid/content/Intent;
 }
 
+public abstract interface class com/freeletics/khonshu/navigation/NavDestination {
+}
+
 public class com/freeletics/khonshu/navigation/NavEventNavigator : com/freeletics/khonshu/navigation/ActivityResultNavigator, com/freeletics/khonshu/navigation/BackInterceptor, com/freeletics/khonshu/navigation/Navigator, com/freeletics/khonshu/navigation/ResultNavigator {
+	public static final field $stable I
 	public fun <init> ()V
 	public fun backPresses ()Lkotlinx/coroutines/flow/Flow;
 	public fun backPresses (Ljava/lang/Object;)Lkotlinx/coroutines/flow/Flow;
@@ -65,10 +80,12 @@ public abstract interface class com/freeletics/khonshu/navigation/NavRoute : com
 }
 
 public final class com/freeletics/khonshu/navigation/NavigationResultRequest : com/freeletics/khonshu/navigation/ResultOwner {
+	public static final field $stable I
 	public final fun getKey ()Lcom/freeletics/khonshu/navigation/NavigationResultRequest$Key;
 }
 
 public final class com/freeletics/khonshu/navigation/NavigationResultRequest$Key : android/os/Parcelable {
+	public static final field $stable I
 	public static final field CREATOR Landroid/os/Parcelable$Creator;
 	public fun describeContents ()I
 	public fun equals (Ljava/lang/Object;)Z
@@ -83,6 +100,10 @@ public final class com/freeletics/khonshu/navigation/NavigationResultRequest$Key
 	public synthetic fun createFromParcel (Landroid/os/Parcel;)Ljava/lang/Object;
 	public final fun newArray (I)[Lcom/freeletics/khonshu/navigation/NavigationResultRequest$Key;
 	public synthetic fun newArray (I)[Ljava/lang/Object;
+}
+
+public final class com/freeletics/khonshu/navigation/NavigationSetupKt {
+	public static final fun NavigationSetup (Lcom/freeletics/khonshu/navigation/NavEventNavigator;Landroidx/compose/runtime/Composer;I)V
 }
 
 public abstract interface class com/freeletics/khonshu/navigation/Navigator {
@@ -100,6 +121,7 @@ public final class com/freeletics/khonshu/navigation/Navigator$Companion {
 }
 
 public final class com/freeletics/khonshu/navigation/PermissionsResultRequest : com/freeletics/khonshu/navigation/ContractResultOwner {
+	public static final field $stable I
 	public synthetic fun getContract ()Landroidx/activity/result/contract/ActivityResultContract;
 }
 
@@ -107,6 +129,7 @@ public abstract interface class com/freeletics/khonshu/navigation/PermissionsRes
 }
 
 public final class com/freeletics/khonshu/navigation/PermissionsResultRequest$PermissionResult$Denied : com/freeletics/khonshu/navigation/PermissionsResultRequest$PermissionResult {
+	public static final field $stable I
 	public fun <init> (Z)V
 	public fun equals (Ljava/lang/Object;)Z
 	public final fun getShouldShowRationale ()Z
@@ -115,6 +138,7 @@ public final class com/freeletics/khonshu/navigation/PermissionsResultRequest$Pe
 }
 
 public final class com/freeletics/khonshu/navigation/PermissionsResultRequest$PermissionResult$Granted : com/freeletics/khonshu/navigation/PermissionsResultRequest$PermissionResult {
+	public static final field $stable I
 	public static final field INSTANCE Lcom/freeletics/khonshu/navigation/PermissionsResultRequest$PermissionResult$Granted;
 	public fun equals (Ljava/lang/Object;)Z
 	public fun hashCode ()I
@@ -126,6 +150,7 @@ public abstract interface class com/freeletics/khonshu/navigation/ResultNavigato
 }
 
 public abstract class com/freeletics/khonshu/navigation/ResultOwner {
+	public static final field $stable I
 	public final fun getResults ()Lkotlinx/coroutines/flow/Flow;
 }
 
@@ -146,6 +171,7 @@ public final class com/freeletics/khonshu/navigation/deeplinks/AndroidDeepLinkEx
 }
 
 public final class com/freeletics/khonshu/navigation/deeplinks/DeepLink {
+	public static final field $stable I
 	public fun equals (Ljava/lang/Object;)Z
 	public fun hashCode ()I
 	public fun toString ()Ljava/lang/String;
@@ -199,17 +225,6 @@ public final class com/freeletics/khonshu/navigation/deeplinks/DeepLinkParserKt 
 public final class com/freeletics/khonshu/navigation/internal/DestinationIdKt {
 }
 
-public final class com/freeletics/khonshu/navigation/internal/InitialValue$Creator : android/os/Parcelable$Creator {
-	public fun <init> ()V
-	public final fun createFromParcel (Landroid/os/Parcel;)Lcom/freeletics/khonshu/navigation/internal/InitialValue;
-	public synthetic fun createFromParcel (Landroid/os/Parcel;)Ljava/lang/Object;
-	public final fun newArray (I)[Lcom/freeletics/khonshu/navigation/internal/InitialValue;
-	public synthetic fun newArray (I)[Ljava/lang/Object;
-}
-
 public abstract interface annotation class com/freeletics/khonshu/navigation/internal/InternalNavigationApi : java/lang/annotation/Annotation {
-}
-
-public final class com/freeletics/khonshu/navigation/internal/NavigationSetupKt {
 }
