+++ navigation/src/androidMain/kotlin/com/freeletics/khonshu/navigation/NavEventNavigator.kt
@@ -164,7 +164,7 @@ public open class NavEventNavigator : Navigator, ResultNavigator, ActivityResult
      *
      * For permission requests prefer using [registerForPermissionsResult] instead.
      *
-     * Note: You must call this before `NavigationSetup` is called with this navigator."
+     * Note: You must call this before [NavigationSetup] is called with this navigator."
      */
     protected fun <I, O> registerForActivityResult(
         contract: ActivityResultContract<I, O>,
@@ -187,7 +187,7 @@ public open class NavEventNavigator : Navigator, ResultNavigator, ActivityResult
      * a `PermissionResult` instead of a `boolean. See `[PermissionsResultRequest.PermissionResult]`
      * for more information.
      *
-     * Note: You must call this before `NavigationSetup` is called with this navigator."
+     * Note: You must call this before [NavigationSetup] is called with this navigator."
      */
     protected fun registerForPermissionsResult(): PermissionsResultRequest {
         checkAllowedToAddRequests()
@@ -203,7 +203,7 @@ public open class NavEventNavigator : Navigator, ResultNavigator, ActivityResult
      * The returned [NavigationResultRequest] has a [NavigationResultRequest.Key]. This `key` should
      * be passed to the target destination which can then use it to call [deliverNavigationResult].
      *
-     * Note: You must call this before `NavigationSetup` is called with this navigator."
+     * Note: You must call this before [NavigationSetup] is called with this navigator."
      */
     protected inline fun <reified T : BaseRoute, reified O : Parcelable> registerForNavigationResult():
         NavigationResultRequest<O> {
